backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
google-genai==1.0.0
qrcode[pil]==8.0
Pillow==11.0.0
pdf2image==1.17.0
python-dotenv==1.0.1